requirements.txt
gunicorn==19.0.0
line-bot-sdk
django
BeautifulSoup4